rename: `PatientFilterV2` -> `PatientFilter`
